refactor!: rewrite Vercel API with typescript
- add Vercel node package
