+++ api/cards/stats.ts
@@ -1,30 +1,26 @@
-const {
-    getCommitsLanguageSVGWithThemeName,
-} = require('../../src/cards/most-commit-lauguage-card');
-const { changToNextGitHubToken } = require('../utils/github-token-updater');
-const { getErrorMsgCard } = require('../utils/error-card');
+import {getStatsSVGWithThemeName} from '../../src/cards/stats-card';
+import {changToNextGitHubToken} from '../utils/github-token-updater';
+import {getErrorMsgCard} from '../utils/error-card';
+import type {VercelRequest, VercelResponse} from '@vercel/node';
 
-module.exports = async (req, res) => {
-    const { username, theme="default" } = req.query;
+export default async (req: VercelRequest, res: VercelResponse) => {
+    const {username, theme = 'default'} = req.query;
     try {
         let tokenIndex = 0;
         while (true) {
             try {
-                const cardSVG = await getCommitsLanguageSVGWithThemeName(
-                    username,
-                    theme
-                );
+                const cardSVG = await getStatsSVGWithThemeName(username, theme);
                 res.setHeader('Content-Type', 'image/svg+xml');
                 res.send(cardSVG);
                 return;
-            } catch (err) {
+            } catch (err: any) {
                 console.log(err.message);
                 // We update github token and try again, until getNextGitHubToken throw an Error
                 changToNextGitHubToken(tokenIndex);
                 tokenIndex += 1;
             }
         }
-    } catch (err) {
+    } catch (err: any) {
         console.log(err);
         res.send(getErrorMsgCard(err.message, theme));
     }
