+++ api/utils/github-token-updater.ts
@@ -1,12 +1,8 @@
-const changToNextGitHubToken = function (currentIndex) {
+export const changToNextGitHubToken = function (currentIndex: number) {
     const tokenName = `GITHUB_TOKEN_${currentIndex + 1}`;
     console.log(`Change to ${tokenName}`);
     process.env.GITHUB_TOKEN = process.env[tokenName];
     if (!process.env.GITHUB_TOKEN) {
         throw new Error('No more GITHUB_TOKEN can be used');
     }
-};
-
-module.exports = {
-    changToNextGitHubToken,
 };
